Crate fe_analyzer

Fe semantic analysis.

This library is used to analyze the semantics of a given Fe AST. It detects any semantic errors within a given AST and produces a Context instance that can be used to query contextual information attributed to AST nodes.

Re-exports

pub use db::AnalyzerDb;

pub use db::TestDb;

Modules

builtins

constants

context

db

display

errors

Semantic errors.

namespace

pattern_analysis

This module includes utility structs and its functions for pattern matching analysis. The algorithm here is based on Warnings for pattern matching

Functions

analyze_ingot

analyze_module