Struct fe_analyzer::context::Context

pub struct Context {
    pub node_ids: Vec<NodeId>,
    pub spans: BTreeMap<NodeId, Span>,
    pub expressions: BTreeMap<NodeId, ExpressionAttributes>,
    pub emits: BTreeMap<NodeId, EventDef>,
    pub functions: BTreeMap<NodeId, FunctionAttributes>,
    pub declarations: BTreeMap<NodeId, FixedSize>,
    pub contracts: BTreeMap<NodeId, ContractAttributes>,
    pub calls: BTreeMap<NodeId, CallType>,
    pub events: BTreeMap<NodeId, EventDef>,
    pub type_descs: BTreeMap<NodeId, Type>,
    pub module: Option<ModuleAttributes>,
    pub file_id: SourceFileId,
    pub diagnostics: Vec<Diagnostic>,
    // some fields omitted
}

Contains contextual information about a Fe module and can be queried using Spanned AST nodes.

Fields
node_ids: Vec<NodeId>

Node ids in the order they were visited.

spans: BTreeMap<NodeId, Span>

The span of a given node id.

expressions: BTreeMap<NodeId, ExpressionAttributes>emits: BTreeMap<NodeId, EventDef>functions: BTreeMap<NodeId, FunctionAttributes>declarations: BTreeMap<NodeId, FixedSize>contracts: BTreeMap<NodeId, ContractAttributes>calls: BTreeMap<NodeId, CallType>events: BTreeMap<NodeId, EventDef>type_descs: BTreeMap<NodeId, Type>module: Option<ModuleAttributes>file_id: SourceFileIddiagnostics: Vec<Diagnostic>
Implementations

impl Context

pub fn new_shared(file_id: SourceFileId) -> Shared<Self>

pub fn new(file_id: SourceFileId) -> Self

pub fn add_expression(
&mut self,
node: &Node<Expr>,
attributes: ExpressionAttributes
)

Attribute contextual information to an expression node.

Panics

Panics if an entry already exists for the node id.

pub fn update_expression(
&mut self,
node: &Node<Expr>,
attributes: ExpressionAttributes
)

Update the expression attributes.

Panics

Panics if an entry does not already exist for the node id.

pub fn get_expression<T: Into<NodeId>>(
&self,
node_id: T
) -> Option<&ExpressionAttributes>

Get information that has been attributed to an expression node.

pub fn add_emit(&mut self, node: &Node<FuncStmt>, event: EventDef)

Attribute contextual information to an emit statement node.

Panics

Panics if an entry already exists for the node id.

pub fn get_emit<T: Into<NodeId>>(&self, node_id: T) -> Option<&EventDef>

Get information that has been attributed to an emit statement node.

pub fn add_function(
&mut self,
node: &Node<ContractStmt>,
attributes: FunctionAttributes
)

Attribute contextual information to a function definition node.

Panics

Panics if an entry already exists for the node id.

pub fn get_function<T: Into<NodeId>>(
&self,
node_id: T
) -> Option<&FunctionAttributes>

Get information that has been attributed to a function definition node.

pub fn add_declaration(&mut self, node: &Node<FuncStmt>, typ: FixedSize)

Attribute contextual information to a declaration node.

Panics

Panics if an entry already exists for the node id.

pub fn get_declaration<T: Into<NodeId>>(&self, node_id: T) -> Option<&FixedSize>

Get information that has been attributed to a declaration node.

pub fn add_contract(
&mut self,
node: &Node<ModuleStmt>,
attributes: ContractAttributes
)

Attribute contextual information to a contract definition node.

Panics

Panics if an entry already exists for the node id.

pub fn get_contract<T: Into<NodeId>>(
&self,
node_id: T
) -> Option<&ContractAttributes>

Get information that has been attributed to a contract definition node.

pub fn add_call(&mut self, node: &Node<Expr>, call_type: CallType)

Attribute contextual information to a call expression node.

Panics

Panics if an entry already exists for the node id.

pub fn get_call<T: Into<NodeId>>(&self, node_id: T) -> Option<&CallType>

Get information that has been attributed to a call expression node.

pub fn add_event(&mut self, node: &Node<ContractStmt>, event: EventDef)

Attribute contextual information to an event definition node.

Panics

Panics if an entry already exists for the node id.

pub fn get_event<T: Into<NodeId>>(&self, node_id: T) -> Option<&EventDef>

Get information that has been attributed to an event definition node.

pub fn add_type_desc(&mut self, node: &Node<TypeDesc>, typ: Type)

Attribute contextual information to a type description node.

Panics

Panics if an entry already exists for the node id.

pub fn get_type_desc<T: Into<NodeId>>(&self, node_id: T) -> Option<&Type>

Get information that has been attributed to a type description node.

pub fn set_module(&mut self, attributes: ModuleAttributes)

Attribute contextual information to the module.

pub fn get_module(&self) -> Option<&ModuleAttributes>

Get information that has been attributed to the module.

pub fn get_spanned_expressions(&self) -> Vec<(Span, ExpressionAttributes)>

Get the span and attributes of all expressions.

pub fn get_spanned_emits(&self) -> Vec<(Span, EventDef)>

Get the span and attributes of all emits.

pub fn get_spanned_functions(&self) -> Vec<(Span, FunctionAttributes)>

Get the span and attributes of all functions.

pub fn get_spanned_declarations(&self) -> Vec<(Span, FixedSize)>

Get the span and attributes of all declarations.

pub fn get_spanned_contracts(&self) -> Vec<(Span, ContractAttributes)>

Get the span and attributes of all contracts.

pub fn get_spanned_calls(&self) -> Vec<(Span, CallType)>

Get the span and attributes of all calls.

pub fn get_spanned_events(&self) -> Vec<(Span, EventDef)>

Get the span and attributes of all events.

pub fn get_spanned_type_descs(&self) -> Vec<(Span, Type)>

Get the span and attributes of all type descs.

pub fn error(
&mut self,
message: impl Into<String>,
label_span: Span,
label: impl Into<String>
)

pub fn type_error(
&mut self,
message: impl Into<String>,
span: Span,
expected: impl Display,
actual: impl Display
)

pub fn not_yet_implemented(&mut self, feature: impl Display, span: Span)

pub fn fancy_error(
&mut self,
message: impl Into<String>,
labels: Vec<Label>,
notes: Vec<String>
)

pub fn make_unique_name(&mut self, name: &str) -> String

Makes a unique name from the given name, keeping it as readable as possible.