Trait fe_analyzer::context::AnalyzerContext

pub trait AnalyzerContext {

    fn resolve_name(
        &self, 
        name: &str
    ) -> Result<Option<NamedThing>, IncompleteItem>;

    fn resolve_path(&mut self, path: &Path) -> Option<NamedThing>;

    fn add_diagnostic(&mut self, diag: Diagnostic);

    fn db(&self) -> &dyn AnalyzerDb;

    fn add_expression(
        &self, 
        node: &Node<Expr>, 
        attributes: ExpressionAttributes
    );

    fn update_expression(
        &self, 
        node: &Node<Expr>, 
        attributes: ExpressionAttributes
    );

    fn expr_typ(&self, expr: &Node<Expr>) -> Type;

    fn add_constant(
        &self, 
        name: &Node<SmolStr>, 
        expr: &Node<Expr>, 
        value: Constant
    );

    fn constant_value_by_name(
        &self, 
        name: &SmolStr
    ) -> Result<Option<Constant>, IncompleteItem>;

    fn parent(&self) -> Item;

    fn parent_function(&self) -> FunctionId;

    fn add_call(&self, node: &Node<Expr>, call_type: CallType);

    fn add_string(&self, str_lit: SmolStr);

    fn is_in_function(&self) -> bool;

    fn inherits_type(&self, typ: BlockScopeType) -> bool;

    fn error(
        &mut self, 
        message: &str, 
        label_span: Span, 
        label: &str
    ) -> DiagnosticVoucher { ... }

    fn root_item(&self) -> Item { ... }

    fn type_error(
        &mut self, 
        message: &str, 
        span: Span, 
        expected: &dyn Display, 
        actual: &dyn Display
    ) -> DiagnosticVoucher { ... }

    fn not_yet_implemented(
        &mut self, 
        feature: &str, 
        span: Span
    ) -> DiagnosticVoucher { ... }

    fn fancy_error(
        &mut self, 
        message: &str, 
        labels: Vec<Label>, 
        notes: Vec<String>
    ) -> DiagnosticVoucher { ... }

    fn duplicate_name_error(
        &mut self, 
        message: &str, 
        name: &str, 
        original: Span, 
        duplicate: Span
    ) -> DiagnosticVoucher { ... }

    fn name_conflict_error(
        &mut self, 
        name_kind: &str, 
        name: &str, 
        original: &NamedThing, 
        original_span: Option<Span>, 
        duplicate_span: Span
    ) -> DiagnosticVoucher { ... }

    fn register_diag(&mut self, diag: Diagnostic) -> DiagnosticVoucher { ... }

}

Required methods

fn resolve_name(&self, name: &str) -> Result<Option<NamedThing>, IncompleteItem>

fn resolve_path(&mut self, path: &Path) -> Option<NamedThing>

fn add_diagnostic(&mut self, diag: Diagnostic)

fn db(&self) -> &dyn AnalyzerDb

fn add_expression(&self, node: &Node<Expr>, attributes: ExpressionAttributes)

Attribute contextual information to an expression node.

Panics

Panics if an entry already exists for the node id.

fn update_expression(&self, node: &Node<Expr>, attributes: ExpressionAttributes)

Update the expression attributes.

Panics

Panics if an entry does not already exist for the node id.

fn expr_typ(&self, expr: &Node<Expr>) -> Type

Returns a type of an expression.

Panics

Panics if type analysis is not performed for an expr.

fn add_constant(&self, name: &Node<SmolStr>, expr: &Node<Expr>, value: Constant)

Add evaluated constant value in a constant declaration to the context.

fn constant_value_by_name(
&self,
name: &SmolStr
) -> Result<Option<Constant>, IncompleteItem>

Returns constant value from variable name.

fn parent(&self) -> Item

Returns an item enclosing current context.

Example

contract Foo:
    fn foo():
       if ...:
           ...
       else:
           ...

If the context is in then block, then this function returns Item::Function(..).

fn parent_function(&self) -> FunctionId

Returns a function id that encloses a context.

Panics

Panics if a context is not in a function. Use Self::is_in_function to determine whether a context is in a function.

fn add_call(&self, node: &Node<Expr>, call_type: CallType)

Panics

Panics if a context is not in a function. Use Self::is_in_function to determine whether a context is in a function.

fn add_string(&self, str_lit: SmolStr)

Store string literal to the current context.

Panics

Panics if a context is not in a function. Use Self::is_in_function to determine whether a context is in a function.

fn is_in_function(&self) -> bool

Returns true if the context is in function scope.

fn inherits_type(&self, typ: BlockScopeType) -> bool

Returns true if the scope or any of its parents is of the given type.

Provided methods

fn error(
&mut self,
message: &str,
label_span: Span,
label: &str
) -> DiagnosticVoucher

fn root_item(&self) -> Item

Returns a non-function item that encloses a context.

Example

contract Foo:
    fn foo():
       if ...:
           ...
       else:
           ...

If the context is in then block, then this function returns Item::Type(TypeDef::Contract(..)).

fn type_error(
&mut self,
message: &str,
span: Span,
expected: &dyn Display,
actual: &dyn Display
) -> DiagnosticVoucher

fn not_yet_implemented(
&mut self,
feature: &str,
span: Span
) -> DiagnosticVoucher

fn fancy_error(
&mut self,
message: &str,
labels: Vec<Label>,
notes: Vec<String>
) -> DiagnosticVoucher

fn duplicate_name_error(
&mut self,
message: &str,
name: &str,
original: Span,
duplicate: Span
) -> DiagnosticVoucher

fn name_conflict_error(
&mut self,
name_kind: &str,
name: &str,
original: &NamedThing,
original_span: Option<Span>,
duplicate_span: Span
) -> DiagnosticVoucher

fn register_diag(&mut self, diag: Diagnostic) -> DiagnosticVoucher