fe_analyzer::db

Struct InternIngotQuery

pub struct InternIngotQuery;

Implementations

impl InternIngotQuery

pub fn in_db(self, db: &dyn AnalyzerDb) -> QueryTable<'_, Self>

Get access to extra methods pertaining to this query. For example, you can use this to run the GC (sweep) across a single input. You can also use it to invoke this query, though it’s more common to use the trait method on the database itself.

impl InternIngotQuery

pub fn in_db_mut(self, db: &mut dyn AnalyzerDb) -> QueryTableMut<'_, Self>

Like in_db, but gives access to methods for setting the value of an input. Not applicable to derived queries.

Threads, cancellation, and blocking

Mutating the value of a query cannot be done while there are still other queries executing. If you are using your database within a single thread, this is not a problem: you only have &self access to the database, but this method requires &mut self.

However, if you have used snapshot to create other threads, then attempts to set will block the current thread until those snapshots are dropped (usually when those threads complete). This also implies that if you create a snapshot but do not send it to another thread, then invoking set will deadlock.

Before blocking, the thread that is attempting to set will also set a cancellation flag. In the threads operating on snapshots, you can use the is_current_revision_canceled method to check for this flag and bring those operations to a close, thus allowing the set to succeed. Ignoring this flag may lead to “starvation”, meaning that the thread attempting to set has to wait a long, long time. =)