Trait fe_analyzer::db::AnalyzerDb

pub trait AnalyzerDb: Database + HasQueryGroup<AnalyzerDbStorage> + SourceDb + Upcast<dyn SourceDb> + UpcastMut<dyn SourceDb> {

    fn intern_ingot(&self, key0: Rc<Ingot>) -> IngotId;

    fn lookup_intern_ingot(&self, key0: IngotId) -> Rc<Ingot>;

    fn intern_module(&self, key0: Rc<Module>) -> ModuleId;

    fn lookup_intern_module(&self, key0: ModuleId) -> Rc<Module>;

    fn intern_module_const(&self, key0: Rc<ModuleConstant>) -> ModuleConstantId;

    fn lookup_intern_module_const(
        &self, 
        key0: ModuleConstantId
    ) -> Rc<ModuleConstant>;

    fn intern_struct(&self, key0: Rc<Struct>) -> StructId;

    fn lookup_intern_struct(&self, key0: StructId) -> Rc<Struct>;

    fn intern_struct_field(&self, key0: Rc<StructField>) -> StructFieldId;

    fn lookup_intern_struct_field(&self, key0: StructFieldId) -> Rc<StructField>;

    fn intern_type_alias(&self, key0: Rc<TypeAlias>) -> TypeAliasId;

    fn lookup_intern_type_alias(&self, key0: TypeAliasId) -> Rc<TypeAlias>;

    fn intern_contract(&self, key0: Rc<Contract>) -> ContractId;

    fn lookup_intern_contract(&self, key0: ContractId) -> Rc<Contract>;

    fn intern_contract_field(&self, key0: Rc<ContractField>) -> ContractFieldId;

    fn lookup_intern_contract_field(
        &self, 
        key0: ContractFieldId
    ) -> Rc<ContractField>;

    fn intern_function(&self, key0: Rc<Function>) -> FunctionId;

    fn lookup_intern_function(&self, key0: FunctionId) -> Rc<Function>;

    fn intern_event(&self, key0: Rc<Event>) -> EventId;

    fn lookup_intern_event(&self, key0: EventId) -> Rc<Event>;

    fn ingot_modules(&self, key0: IngotId) -> Rc<[ModuleId]>;

    fn set_ingot_modules(&mut self, key0: IngotId, value__: Rc<[ModuleId]>);

    fn set_ingot_modules_with_durability(
        &mut self, 
        key0: IngotId, 
        value__: Rc<[ModuleId]>, 
        durability__: Durability
    );

    fn ingot_external_ingots(
        &self, 
        key0: IngotId
    ) -> Rc<IndexMap<SmolStr, IngotId>>;

    fn set_ingot_external_ingots(
        &mut self, 
        key0: IngotId, 
        value__: Rc<IndexMap<SmolStr, IngotId>>
    );

    fn set_ingot_external_ingots_with_durability(
        &mut self, 
        key0: IngotId, 
        value__: Rc<IndexMap<SmolStr, IngotId>>, 
        durability__: Durability
    );

    fn ingot_root_module(&self, key0: IngotId) -> Option<ModuleId>;

    fn module_file_path(&self, key0: ModuleId) -> SmolStr;

    fn module_parse(&self, key0: ModuleId) -> Analysis<Rc<Module>>;

    fn module_is_incomplete(&self, key0: ModuleId) -> bool;

    fn module_all_items(&self, key0: ModuleId) -> Rc<[Item]>;

    fn module_item_map(
        &self, 
        key0: ModuleId
    ) -> Analysis<Rc<IndexMap<SmolStr, Item>>>;

    fn module_contracts(&self, key0: ModuleId) -> Rc<[ContractId]>;

    fn module_structs(&self, key0: ModuleId) -> Rc<[StructId]>;

    fn module_constants(&self, key0: ModuleId) -> Rc<Vec<ModuleConstantId>>;

    fn module_used_item_map(
        &self, 
        key0: ModuleId
    ) -> Analysis<Rc<IndexMap<SmolStr, (Span, Item)>>>;

    fn module_parent_module(&self, key0: ModuleId) -> Option<ModuleId>;

    fn module_submodules(&self, key0: ModuleId) -> Rc<[ModuleId]>;

    fn module_constant_type(
        &self, 
        key0: ModuleConstantId
    ) -> Analysis<Result<Type, TypeError>>;

    fn module_constant_value(
        &self, 
        key0: ModuleConstantId
    ) -> Analysis<Result<Constant, ConstEvalError>>;

    fn contract_all_functions(&self, key0: ContractId) -> Rc<[FunctionId]>;

    fn contract_function_map(
        &self, 
        key0: ContractId
    ) -> Analysis<Rc<IndexMap<SmolStr, FunctionId>>>;

    fn contract_public_function_map(
        &self, 
        key0: ContractId
    ) -> Rc<IndexMap<SmolStr, FunctionId>>;

    fn contract_init_function(
        &self, 
        key0: ContractId
    ) -> Analysis<Option<FunctionId>>;

    fn contract_call_function(
        &self, 
        key0: ContractId
    ) -> Analysis<Option<FunctionId>>;

    fn contract_all_events(&self, key0: ContractId) -> Rc<[EventId]>;

    fn contract_event_map(
        &self, 
        key0: ContractId
    ) -> Analysis<Rc<IndexMap<SmolStr, EventId>>>;

    fn contract_all_fields(&self, key0: ContractId) -> Rc<[ContractFieldId]>;

    fn contract_field_map(
        &self, 
        key0: ContractId
    ) -> Analysis<Rc<IndexMap<SmolStr, ContractFieldId>>>;

    fn contract_field_type(
        &self, 
        key0: ContractFieldId
    ) -> Analysis<Result<Type, TypeError>>;

    fn contract_dependency_graph(&self, key0: ContractId) -> DepGraphWrapper;

    fn contract_runtime_dependency_graph(
        &self, 
        key0: ContractId
    ) -> DepGraphWrapper;

    fn function_signature(
        &self, 
        key0: FunctionId
    ) -> Analysis<Rc<FunctionSignature>>;

    fn function_body(&self, key0: FunctionId) -> Analysis<Rc<FunctionBody>>;

    fn function_dependency_graph(&self, key0: FunctionId) -> DepGraphWrapper;

    fn struct_type(&self, key0: StructId) -> Rc<Struct>;

    fn struct_all_fields(&self, key0: StructId) -> Rc<[StructFieldId]>;

    fn struct_field_map(
        &self, 
        key0: StructId
    ) -> Analysis<Rc<IndexMap<SmolStr, StructFieldId>>>;

    fn struct_field_type(
        &self, 
        key0: StructFieldId
    ) -> Analysis<Result<FixedSize, TypeError>>;

    fn struct_all_functions(&self, key0: StructId) -> Rc<[FunctionId]>;

    fn struct_function_map(
        &self, 
        key0: StructId
    ) -> Analysis<Rc<IndexMap<SmolStr, FunctionId>>>;

    fn struct_dependency_graph(&self, key0: StructId) -> DepGraphWrapper;

    fn event_type(&self, key0: EventId) -> Analysis<Rc<Event>>;

    fn type_alias_type(
        &self, 
        key0: TypeAliasId
    ) -> Analysis<Result<Type, TypeError>>;

}

Required methods

fn intern_ingot(&self, key0: Rc<Ingot>) -> IngotId

fn lookup_intern_ingot(&self, key0: IngotId) -> Rc<Ingot>

fn intern_module(&self, key0: Rc<Module>) -> ModuleId

fn lookup_intern_module(&self, key0: ModuleId) -> Rc<Module>

fn intern_module_const(&self, key0: Rc<ModuleConstant>) -> ModuleConstantId

fn lookup_intern_module_const(
&self,
key0: ModuleConstantId
) -> Rc<ModuleConstant>

fn intern_struct(&self, key0: Rc<Struct>) -> StructId

fn lookup_intern_struct(&self, key0: StructId) -> Rc<Struct>

fn intern_struct_field(&self, key0: Rc<StructField>) -> StructFieldId

fn lookup_intern_struct_field(&self, key0: StructFieldId) -> Rc<StructField>

fn intern_type_alias(&self, key0: Rc<TypeAlias>) -> TypeAliasId

fn lookup_intern_type_alias(&self, key0: TypeAliasId) -> Rc<TypeAlias>

fn intern_contract(&self, key0: Rc<Contract>) -> ContractId

fn lookup_intern_contract(&self, key0: ContractId) -> Rc<Contract>

fn intern_contract_field(&self, key0: Rc<ContractField>) -> ContractFieldId

fn lookup_intern_contract_field(
&self,
key0: ContractFieldId
) -> Rc<ContractField>

fn intern_function(&self, key0: Rc<Function>) -> FunctionId

fn lookup_intern_function(&self, key0: FunctionId) -> Rc<Function>

fn intern_event(&self, key0: Rc<Event>) -> EventId

fn lookup_intern_event(&self, key0: EventId) -> Rc<Event>

fn ingot_modules(&self, key0: IngotId) -> Rc<[ModuleId]>

fn set_ingot_modules(&mut self, key0: IngotId, value__: Rc<[ModuleId]>)

Set the value of the ingot_modules input.

See ingot_modules for details.

Note: Setting values will trigger cancellation of any ongoing queries; this method blocks until those queries have been cancelled.

fn set_ingot_modules_with_durability(
&mut self,
key0: IngotId,
value__: Rc<[ModuleId]>,
durability__: Durability
)

Set the value of the ingot_modules input and promise that its value will never change again.

See ingot_modules for details.

Note: Setting values will trigger cancellation of any ongoing queries; this method blocks until those queries have been cancelled.

fn ingot_external_ingots(&self, key0: IngotId) -> Rc<IndexMap<SmolStr, IngotId>>

fn set_ingot_external_ingots(
&mut self,
key0: IngotId,
value__: Rc<IndexMap<SmolStr, IngotId>>
)

Set the value of the ingot_external_ingots input.

See ingot_external_ingots for details.

Note: Setting values will trigger cancellation of any ongoing queries; this method blocks until those queries have been cancelled.

fn set_ingot_external_ingots_with_durability(
&mut self,
key0: IngotId,
value__: Rc<IndexMap<SmolStr, IngotId>>,
durability__: Durability
)

Set the value of the ingot_external_ingots input and promise that its value will never change again.

See ingot_external_ingots for details.

Note: Setting values will trigger cancellation of any ongoing queries; this method blocks until those queries have been cancelled.

fn ingot_root_module(&self, key0: IngotId) -> Option<ModuleId>

fn module_file_path(&self, key0: ModuleId) -> SmolStr

fn module_parse(&self, key0: ModuleId) -> Analysis<Rc<Module>>

fn module_is_incomplete(&self, key0: ModuleId) -> bool

fn module_all_items(&self, key0: ModuleId) -> Rc<[Item]>

fn module_item_map(
&self,
key0: ModuleId
) -> Analysis<Rc<IndexMap<SmolStr, Item>>>

fn module_contracts(&self, key0: ModuleId) -> Rc<[ContractId]>

fn module_structs(&self, key0: ModuleId) -> Rc<[StructId]>

fn module_constants(&self, key0: ModuleId) -> Rc<Vec<ModuleConstantId>>

fn module_used_item_map(
&self,
key0: ModuleId
) -> Analysis<Rc<IndexMap<SmolStr, (Span, Item)>>>

fn module_parent_module(&self, key0: ModuleId) -> Option<ModuleId>

fn module_submodules(&self, key0: ModuleId) -> Rc<[ModuleId]>

fn module_constant_type(
&self,
key0: ModuleConstantId
) -> Analysis<Result<Type, TypeError>>

fn module_constant_value(
&self,
key0: ModuleConstantId
) -> Analysis<Result<Constant, ConstEvalError>>

fn contract_all_functions(&self, key0: ContractId) -> Rc<[FunctionId]>

fn contract_function_map(
&self,
key0: ContractId
) -> Analysis<Rc<IndexMap<SmolStr, FunctionId>>>

fn contract_public_function_map(
&self,
key0: ContractId
) -> Rc<IndexMap<SmolStr, FunctionId>>

fn contract_init_function(
&self,
key0: ContractId
) -> Analysis<Option<FunctionId>>

fn contract_call_function(
&self,
key0: ContractId
) -> Analysis<Option<FunctionId>>

fn contract_all_events(&self, key0: ContractId) -> Rc<[EventId]>

fn contract_event_map(
&self,
key0: ContractId
) -> Analysis<Rc<IndexMap<SmolStr, EventId>>>

fn contract_all_fields(&self, key0: ContractId) -> Rc<[ContractFieldId]>

fn contract_field_map(
&self,
key0: ContractId
) -> Analysis<Rc<IndexMap<SmolStr, ContractFieldId>>>

fn contract_field_type(
&self,
key0: ContractFieldId
) -> Analysis<Result<Type, TypeError>>

fn contract_dependency_graph(&self, key0: ContractId) -> DepGraphWrapper

fn contract_runtime_dependency_graph(&self, key0: ContractId) -> DepGraphWrapper

fn function_signature(
&self,
key0: FunctionId
) -> Analysis<Rc<FunctionSignature>>

fn function_body(&self, key0: FunctionId) -> Analysis<Rc<FunctionBody>>

fn function_dependency_graph(&self, key0: FunctionId) -> DepGraphWrapper

fn struct_type(&self, key0: StructId) -> Rc<Struct>

fn struct_all_fields(&self, key0: StructId) -> Rc<[StructFieldId]>

fn struct_field_map(
&self,
key0: StructId
) -> Analysis<Rc<IndexMap<SmolStr, StructFieldId>>>

fn struct_field_type(
&self,
key0: StructFieldId
) -> Analysis<Result<FixedSize, TypeError>>

fn struct_all_functions(&self, key0: StructId) -> Rc<[FunctionId]>

fn struct_function_map(
&self,
key0: StructId
) -> Analysis<Rc<IndexMap<SmolStr, FunctionId>>>

fn struct_dependency_graph(&self, key0: StructId) -> DepGraphWrapper

fn event_type(&self, key0: EventId) -> Analysis<Rc<Event>>

fn type_alias_type(
&self,
key0: TypeAliasId
) -> Analysis<Result<Type, TypeError>>