Struct fe_analyzer::namespace::scopes::ContractScope

pub struct ContractScope {
    pub name: String,
    pub parent: Shared<ModuleScope>,
    pub interface: Vec<String>,
    pub event_defs: BTreeMap<String, EventDef>,
    pub field_defs: BTreeMap<String, ContractFieldDef>,
    pub function_defs: BTreeMap<String, ContractFunctionDef>,
    pub list_expressions: BTreeSet<Array>,
    pub string_defs: BTreeSet<String>,
    pub created_contracts: BTreeSet<String>,
    // some fields omitted
}

Fields
name: Stringparent: Shared<ModuleScope>interface: Vec<String>event_defs: BTreeMap<String, EventDef>field_defs: BTreeMap<String, ContractFieldDef>function_defs: BTreeMap<String, ContractFunctionDef>list_expressions: BTreeSet<Array>string_defs: BTreeSet<String>created_contracts: BTreeSet<String>
Implementations

impl ContractScope

pub fn new(name: &str, parent: Shared<ModuleScope>) -> Shared<Self>

pub fn module_scope(&self) -> Shared<ModuleScope>

Return the module scope that the contract scope inherits from

pub fn get_module_type_defs<B, F: FnMut(&Type) -> Option<B>>(
&self,
predicate: F
) -> Vec<B>

Filter module scope for type definitions that match the given predicate

pub fn event_def(&self, name: &str) -> Option<EventDef>

Lookup contract event definition by its name.

pub fn field_def(&self, name: &str) -> Option<ContractFieldDef>

Lookup contract field definition by its name.

pub fn function_def(&self, name: &str) -> Option<ContractFunctionDef>

Lookup contract function definition by its name.

pub fn add_field(&mut self, name: &str, typ: Type) -> Result<(), AlreadyDefined>

Add a contract field definition to the scope.

pub fn add_function(
&mut self,
name: &str,
is_public: bool,
params: Vec<(String, FixedSize)>,
return_type: FixedSize,
scope: Shared<BlockScope>
) -> Result<&ContractFunctionDef, AlreadyDefined>

Add a function definition to the scope.

pub fn add_event(
&mut self,
name: &str,
event: EventDef
) -> Result<(), AlreadyDefined>

Add an event definition to the scope.

pub fn add_string(&mut self, value: &str)

Add a static string definition to the scope.

pub fn add_created_contract(&mut self, name: &str)

Add the name of another contract that has been created within this contract.

pub fn add_used_list_expression(&mut self, typ: Array)

Add the array type of a list expression that was used within the contract.