Module fe_analyzer::namespace::types

Structs

AbiComponent
Single component of a tuple.

AbiUintSize
The size of uint element in the ABI encoding.

Array

Contract

FeString

Map

Struct

Tuple

Enums

AbiArraySize
The size of an array.

AbiDecodeLocation
Data can be decoded from memory or calldata.

AbiType
The type of an element in terms of the ABI spec.

Base

FixedSize

Integer

Type

Constants

U256

Traits

AbiEncoding
Information relevant to ABI encoding.

FeSized
The type has a constant size known to the compiler.

SafeNames
Names that can be used to build identifiers without collision.

TypeDowncast

Functions

i256_max

i256_min

u256_max

u256_min