Trait fe_analyzer::namespace::types::AbiEncoding

pub trait AbiEncoding: SafeNames {
    fn abi_json_name(&self) -> String;

    fn abi_selector_name(&self) -> String;

    fn abi_components(&self) -> Vec<AbiComponent>;

    fn abi_type(&self) -> AbiType;
}

Information relevant to ABI encoding.

Required methods

fn abi_json_name(&self) -> String

Name of the type as it appears in the Json ABI.

fn abi_selector_name(&self) -> String

Name of the type as it appears in the selector preimage.

fn abi_components(&self) -> Vec<AbiComponent>

The components of an ABI tuple.

fn abi_type(&self) -> AbiType

The ABI type of a Fe type.

Loading content...