fe_analyzer

Module pattern_analysis

This module includes utility structs and its functions for pattern matching analysis. The algorithm here is based on Warnings for pattern matching

In this module, we assume all types are well-typed, so we can rely on the type information without checking it.

Structs

PatternMatrix

PatternRowVec

SigmaSet

SimplifiedPattern

Enums

ConstructorKind

LiteralConstructor

SimplifiedPatternKind