Module fe_analyzer::errors

Semantic errors.

Structs

AlreadyDefined

Error to be returned from APIs that should reject duplicate definitions

AnalyzerError

CannotMove

Error indicating that a value can not move between memory and storage

ConstEvalError

Error indicating constant evaluation failed.

FatalError

Error to be returned when otherwise no meaningful information can be returned. Can’t be created unless a diagnostic has been emitted, and thus a DiagnosticVoucher has been obtained. (See comment on TypeError)

IncompleteItem

Error returned by ModuleId::resolve_name if the name is not found, and parsing of the module failed. In this case, emitting an error message about failure to resolve the name might be misleading, because the file may in fact contain an item with the given name, somewhere after the syntax error that caused parsing to fail.

NotFixedSize

Error indicating that a [Type] can’t be converted into a [FixedSize]

TypeError

Error indicating that a type is invalid.

Enums

BinaryOperationError

Errors that can result from a binary operation

IndexingError

Errors that can result from indexing

Functions

duplicate_name_error

error

fancy_error

name_conflict_error

not_yet_implemented

type_error