Module fe_analyzer::namespace::items

Structs

Attribute

AttributeId

Contract

ContractField

ContractFieldId

ContractId

DepGraphWrapper

Enum

EnumId

EnumVariant

EnumVariantId

Function

FunctionId

FunctionSig

FunctionSigId

Impl

ImplId

Ingot

An Ingot is composed of a tree of Modules (set via [AnalyzerDb::set_ingot_module_tree]), and doesn’t have direct knowledge of files.

IngotId

Module

ModuleConstant

ModuleConstantId

ModuleId

Id of a Module, which corresponds to a single Fe source file.

Struct

StructField

StructFieldId

StructId

Trait

TraitId

TypeAlias

TypeAliasId

Enums

DepLocality

DepGraph edge label. “Locality” refers to the deployed state; Local dependencies are those that will be compiled together, while External dependencies will only be reachable via an evm CALL* op.

EnumVariantKind

IngotMode

Item

A named item. This does not include things inside of a function body.

ModuleSource

TypeDef

Traits

DiagnosticSink

Functions

builtin_items

walk_local_dependencies

Type Aliases

DepGraph